Fix missing graphics in the menus
- Initiate osdLeft, osdTop, osdWidth and osdHeight with 0 in
cGeometryManager::cGeometryManager() in order to run through
the function cNopacity::ReloadCaches(void) via
cGeometryManager::SetOSDSize(void) when the skin is started
for the first time.

- cGeometryManager::GeometryChanged(void) is renamed
to cGeometryManager::SetOSDSize(void) and merged to set
osdLeft, osdTop, osdWidth and osdHeight to cOsd::*.

The cache is now created via cNopacity::ReloadCaches(void)
when the skin is started for the first time or when
the geometry or theme is changed.