Simplify cGeometryManager
- Initiate osdLeft, osdTop, osdWidth and osdHeight with 0 in
cGeometryManager::cGeometryManager().

- cGeometryManager::GeometryChanged(void) is renamed
to cGeometryManager::SetOSDSize(void) and merged to set
osdLeft, osdTop, osdWidth and osdHeight to cOsd::*.

The cache is now created via cNopacity::SetOSDSize(void)
when the skin is started for the first time or when
the geometry is changed.

All SetDisplay*() were merged and moved to cGeometryManager::SetGeometry.
So we can delete then all unnecessary funktions.