Optimize Flush in display menu
These changes ensures that only one flush is active when the fade-in thread
is running.
The first Pixmap::Lock() in cNopacityDisplayMenu::Action() takes so much longer
than the following, Thats why the first Pixmap::Lock() was moved before the while
loop. This makes the fade-in softer.
The DetailView thread was eliminated.