Rework Setup
Since the settings are theme-specific and the theme is only known if
this skin is active, changes were not applied if another skin was
active. As a result of this change, the settings of this skin can only
be changed if this skin is active. Otherwise there is a note about it.
In addition, some setup entries have been rearranged.